// src/app/features/pizzas/shared/pizza.enum.ts
export enum Topping {
    Anchovy = 'anchovy',
    Bacon = 'bacon',
    Basil = 'basil',
    Chili = 'chili',
    Mozzarella = 'mozzarella',
    Mushroom = 'mushroom',
    Olive = 'olive',
    Onion = 'onion',
    Pepper = 'pepper',
    Pepperoni = 'pepperoni',
    Sweetcorn = 'sweetcorn',
    Tomato = 'tomato',
}

export enum PizzaSizes {
    Small = 'small',
    Medium = 'medium',
    Large = 'large',
}